Add arcade drive command
