+++ MetalMoose2018CommandBased/src/org/usfirst/frc/team1391/robot/Robot.java
@@ -12,8 +12,8 @@ import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.Scheduler;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import org.usfirst.frc.team1391.robot.commands.ExampleCommand;
-import org.usfirst.frc.team1391.robot.subsystems.ExampleSubsystem;
+import org.usfirst.frc.team1391.robot.commands.ArcadeDrive;
+import org.usfirst.frc.team1391.robot.subsystems.DriveTrain;
 
 /**
  * The VM is configured to automatically run this class, and to call the
@@ -23,8 +23,8 @@ import org.usfirst.frc.team1391.robot.subsystems.ExampleSubsystem;
  * project.
  */
 public class Robot extends TimedRobot {
-	public static final ExampleSubsystem kExampleSubsystem
-			= new ExampleSubsystem();
+	public static final DriveTrain kDriveTrain
+			= new DriveTrain();
 	public static OI m_oi;
 
 	Command m_autonomousCommand;
@@ -37,7 +37,7 @@ public class Robot extends TimedRobot {
 	@Override
 	public void robotInit() {
 		m_oi = new OI();
-		m_chooser.addDefault("Default Auto", new ExampleCommand());
+//		m_chooser.addDefault("Default Auto", new ExampleCommand());
 		// chooser.addObject("My Auto", new MyAutoCommand());
 		SmartDashboard.putData("Auto mode", m_chooser);
 	}
@@ -110,7 +110,6 @@ public class Robot extends TimedRobot {
 	@Override
 	public void teleopPeriodic() {
 		Scheduler.getInstance().run();
-		kExampleSubsystem.diffDrive(OI.stick.getY(), OI.stick.getX());
 	}	
 
 	/**
