+++ MetalMoose2018CommandBased/src/org/usfirst/frc/team1391/robot/commands/ArcadeDrive.java
@@ -1,0 +1,41 @@
+package org.usfirst.frc.team1391.robot.commands;
+
+import edu.wpi.first.wpilibj.command.Command;
+
+import org.usfirst.frc.team1391.robot.OI;
+import org.usfirst.frc.team1391.robot.Robot;
+
+/**
+ *
+ */
+public class ArcadeDrive extends Command {
+
+    public ArcadeDrive() {
+        // Use requires() here to declare subsystem dependencies
+        // eg. requires(chassis);
+    		requires(Robot.kDriveTrain);
+    }
+
+    // Called just before this Command runs the first time
+    protected void initialize() {
+    }
+
+    // Called repeatedly when this Command is scheduled to run
+    protected void execute() {
+		Robot.kDriveTrain.arcadeDrive(OI.stick.getY(), OI.stick.getX());
+    }
+
+    // Make this return true when this Command no longer needs to run execute()
+    protected boolean isFinished() {
+        return false;
+    }
+
+    // Called once after isFinished returns true
+    protected void end() {
+    }
+
+    // Called when another command which requires one or more of the same
+    // subsystems is scheduled to run
+    protected void interrupted() {
+    }
+}
