--- MetalMoose2018CommandBased/src/org/usfirst/frc/team1391/robot/subsystems/ExampleSubsystem.java
@@ -1,36 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package org.usfirst.frc.team1391.robot.subsystems;
-
-import org.usfirst.frc.team1391.robot.RobotMap;
-
-import edu.wpi.first.wpilibj.command.Subsystem;
-import edu.wpi.first.wpilibj.drive.DifferentialDrive;
-import edu.wpi.first.wpilibj.*;
-
-/**
- * An example subsystem.  You can replace me with your own Subsystem.
- */
-public class ExampleSubsystem extends Subsystem {
-	// Put methods for controlling this subsystem
-	// here. Call these from Commands.
-
-	Spark leftMotor = RobotMap.leftMotor;
-	Spark rightMotor = RobotMap.rightMotor;
-	
-	DifferentialDrive drive = new DifferentialDrive(leftMotor,rightMotor);
-	
-	public void initDefaultCommand() {
-		// Set the default command for a subsystem here.
-		// setDefaultCommand(new MySpecialCommand());
-	}
-	
-	public void diffDrive(double d, double e) {
-		drive.arcadeDrive(d, e);
-	}
-}
